Feature vova

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.service.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import ru.skypro.homework.controller.UserController;
 import ru.skypro.homework.dto.AdsComment;
@@ -11,11 +12,14 @@ import ru.skypro.homework.dto.FullAdsDto;
 import ru.skypro.homework.entity.Ads;
 import ru.skypro.homework.entity.Comment;
 import ru.skypro.homework.entity.UserProfile;
+import ru.skypro.homework.exception.AccessDeniedException;
+import ru.skypro.homework.exception.AdsNotFoundException;
 import ru.skypro.homework.exception.CommentNotFoundException;
 import ru.skypro.homework.mapper.AdsMapper;
 import ru.skypro.homework.mapper.CommentMapper;
 import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.repository.CommentRepository;
+import ru.skypro.homework.repository.UserProfileRepository;
 import ru.skypro.homework.service.AdsService;
 
 import java.util.Collection;
@@ -28,10 +32,12 @@ public class AdsServiceImpl implements AdsService {
 
     private final CommentRepository commentRepository;
     private final AdsRepository adsRepository;
+    private final UserProfileRepository userProfileRepository;
 
-    public AdsServiceImpl(CommentRepository commentRepository, AdsRepository adsRepository) {
+    public AdsServiceImpl(CommentRepository commentRepository, AdsRepository adsRepository, UserProfileRepository userProfileRepository) {
         this.commentRepository = commentRepository;
         this.adsRepository = adsRepository;
+        this.userProfileRepository = userProfileRepository;
     }
 
     @Override
@@ -78,8 +84,11 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public void deleteComment(Long adsId, Long commentId) {
+    public void deleteComment(Long adsId, Long commentId, Authentication authentication) {
+
+        checkCommentAccess(adsId, commentId, authentication);
         commentRepository.deleteByAdsIdAndId(adsId, commentId);
+        logger.info("Comment delete successful");
     }
 
     @Override
@@ -91,23 +100,27 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public AdsComment updateAdsComment(Long adsId, Long commentId, AdsComment adsComment) {
+    public AdsComment updateAdsComment(Long adsId, Long commentId, AdsComment adsComment, Authentication authentication) {
 
         Comment oldComment = commentRepository.getByAdsIdAndId(adsId, commentId).get();
         if (oldComment != null) {
 
+            checkCommentAccess(adsId, commentId, authentication);
             Comment newComment = CommentMapper.INSTANCE.adsCommentToComment(adsComment);
-            newComment.setId(oldComment.getId());
+            newComment.setId(commentId);
+            newComment.setAdsId(adsId);
             return mapToAdsComment(commentRepository.save(newComment));
         }
         throw new CommentNotFoundException();
     }
 
     @Override
-    public void removeAds(Long adsId) {
+    public void removeAds(Long adsId, Authentication authentication) {
 
+        checkAdsAccess(adsId, authentication);
         commentRepository.deleteAllByAdsId(adsId);
         adsRepository.deleteAllById(adsId);
+        logger.info("Ads delete successful");
     }
 
     @Override
@@ -119,12 +132,16 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public AdsDto updateAds(AdsDto updatedAds) {
-        Ads ads = AdsMapper.INSTANCE.adsDtoToAds(updatedAds);
+    public AdsDto updateAds(AdsDto updatedAds, Authentication authentication) {
 
-        return AdsMapper
-                .INSTANCE
-                .adsToAdsDto(adsRepository.save(ads));
+        if (adsRepository.existsById(updatedAds.getPk())) {
+            checkAdsAccess(updatedAds.getPk(), authentication);
+            Ads newAds = AdsMapper.INSTANCE.adsDtoToAds(updatedAds);
+            return AdsMapper
+                    .INSTANCE
+                    .adsToAdsDto(adsRepository.save(newAds));
+        }
+        throw  new AdsNotFoundException();
     }
 
     private AdsComment mapToAdsComment(Comment comment) {
@@ -153,5 +170,30 @@ public class AdsServiceImpl implements AdsService {
         fullAds.setAuthorLastName(user.getLastName());
         fullAds.setPhone(user.getPhone());
         return fullAds;
+    }
+
+    private void checkCommentAccess(Long adsId, Long commentId, Authentication authentication) {
+
+        Long userIdFromComments = commentRepository.getUserProfileId(adsId, commentId);
+        Long userIdFromUserProfiles = userProfileRepository.getUserProfileId(authentication.getName());
+        boolean isNotEqualsId = userIdFromComments != userIdFromUserProfiles;
+        checkAccess(isNotEqualsId, authentication);
+    }
+
+    private void checkAdsAccess(Long adsId, Authentication authentication) {
+
+        Long userIdFromAds = adsRepository.getUserProfileId(adsId);
+        Long userIdFromUserProfiles = userProfileRepository.getUserProfileId(authentication.getName());
+        boolean isNotEqualsId = userIdFromAds != userIdFromUserProfiles;
+        checkAccess(isNotEqualsId, authentication);
+    }
+
+    private void checkAccess(boolean isNotEqualsId, Authentication authentication) {
+
+        Boolean noAdminRoots = authentication.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN")) == false;
+
+        if (isNotEqualsId && noAdminRoots) {
+            throw new AccessDeniedException();
+        }
     }
 }
